# app/src/swagger.yaml
openapi: 3.1.0
info:
    title: "Pokémon Catch API"
    version: "1.0.0"
    description: "API for managing Pokémon caught by trainers."
...
paths:
...
    /user:
        get:
            tags:
                - User
            summary: Get all users list
            description: Get all users list
            responses:
                200:
                    $ref: "#/components/responses/User"
        post:
            tags:
                - User
            summary: Create new user
            description: Create new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                password:
                                    type: string
                            required:
                                - name
                                - password
            responses:
                200:
                    $ref: "#/components/responses/UserCreate"
                409:
                    $ref: "#/components/responses/UserCreateConflict"
    /user/{userId}:
        delete:
            tags:
                - User
            summary: Delete user by id
            description: Delete user by id
            parameters:
                - in: path
                  name: userId
                  required: true
                  schema:
                      type: integer
            responses:
                204:
                    description: Deleted succesfully
                    content:
                        application/json:
                            example:
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
    responses:
...
        UserCreate:
            description: Succesfull response
            content:
                application/json:
                    example:
                        - id: 1
                          name: LiptuN
                          password: 937e8d5fbb48bd4949536cd65b8d35c426b80d2f830c5c308e2cdec422ae2244
        UserCreateConflict:
            description: Error response
            content:
                application/json:
                    example:
                        - message: This user name is not available
        User:
            description: Succesfull response
            content:
                application/json:
                    example:
                        - id: 1
                          name: LiptuN
                        - no: 2
                          name: doenutt